refactor(config): use a single LLM wizard entrypoint
